+++ src/srcRcpp.cpp
@@ -1,9 +1,9 @@
 #include <Rcpp.h>
+#include <pbv.h>
+#include<string.h>
 
 // [[Rcpp::depends(pbv)]]
-#include <pbv.h>
 
-#include<string.h>
 using namespace std;
 using namespace Rcpp;
 
@@ -203,6 +203,172 @@ NumericMatrix getRhoMat(int N, NumericVector P, NumericMatrix TCORR) {
   }
   
   return(PCORR);
+  
+}
+
+// [[Rcpp::export]]
+double getBeta0(NumericVector lvec, double popPrev, double tolerance) {
+  
+  double intLow = -10;
+  double intHigh = 10;
+  double B0;
+  double PREV;
+  NumericVector ps(lvec.length());
+  NumericVector nvec(lvec.length());
+  
+  while(abs(intHigh - intLow) > tolerance){
+    
+    B0 = (intLow + intHigh)/2;
+    for (int i = 0; i < lvec.length(); i++) {
+      nvec(i) = lvec(i) + B0;
+    }
+    ps = Rcpp::plogis(nvec);
+    PREV = mean(ps);
+    
+    if (PREV < popPrev) {
+      intLow = B0;
+    } else {
+      intHigh = B0;
+    }
+    
+  }
+  
+  B0 =  (intLow + intHigh)/2;
+  
+  return(B0);
+  
+}
+
+
+
+// [[Rcpp::export]]
+double estAUC(NumericMatrix dmatrix, NumericVector y) {
+  
+  int aucN = 1000000;
+  
+  int N0 = y.length() - sum(y);
+  int N1 = sum(y);
+  
+  NumericVector P1(N1);
+  NumericVector P0(N0);
+  
+  // Obtaining namespace of fastglm package
+  Environment pkg = Environment::namespace_env("fastglm");
+  Function f = pkg["fastglm"];
+  Function p = pkg["predict.fastglm"];
+
+  // model should be include "Named("family", "binomial")", but occasionally 
+  // throws off warning. Using Gaussian works very slightly less well, but
+  // there are no warnings
+  
+  List fit = f(Named("x", dmatrix), Named("y", y)); 
+  NumericVector pred = p(Named("object", fit), Named("newdata", dmatrix));
+  
+  int i1 = 0;
+  int i0 = 0;
+  
+  for (int i = 0; i < y.length(); i++) {
+    if ( y(i) == 1) {
+      P1[i1] = pred(i);
+      i1++;
+    } else {
+      P0[i0] = pred(i);
+      i0++;
+    }
+  } 
+  
+  NumericVector s1 = P1[floor(Rcpp::runif(aucN, 0, N1))];
+  NumericVector s0 = P0[floor(Rcpp::runif(aucN, 0, N0))];
+  
+  double n_greater = 0;
+  
+  for (int i = 0; i < aucN; i++) {
+    if (s1(i) > s0(i)) n_greater++;
+  }
+  
+  double AUC = n_greater/aucN;
+  
+  return(AUC);
+  
+}
+
+// [[Rcpp::export]]
+NumericVector getBeta_auc(NumericMatrix covmat, NumericVector coefs, double auc,
+    double popPrev, double tolerance) {
+  
+  Environment base("package:base");
+  Function mat_Mult = base["%*%"];
+  
+  int N = covmat.nrow();
+  
+  double intLow = 0;
+  double intHigh = 40;
+  
+  double alpha;
+  double B0;
+  double aStat;
+  NumericVector ps(N);
+  NumericVector lvec(N);
+  NumericVector avec(coefs.length());
+  NumericVector avecint(coefs.length() + 1);
+  
+  NumericVector onevec(N); 
+  NumericVector y(N); 
+  
+  NumericVector results(2);
+  
+  for (int i = 0; i < N; i++) {
+    onevec(i) = 1;
+  }
+  
+  NumericMatrix dmatrix = cbind(onevec, covmat);
+  
+  lvec = mat_Mult(covmat, coefs);
+  B0 = getBeta0(lvec, popPrev, tolerance);
+  avecint(0) = B0;
+  
+  aStat = 0;
+  
+  while(abs(aStat - auc) > tolerance) {
+    
+    alpha = (intLow + intHigh)/2;
+    
+    for (int i = 0; i < coefs.length(); i++) {
+      avecint(i+1) = alpha * coefs(i);
+    }
+    
+    lvec = mat_Mult(dmatrix, avecint);
+    ps = Rcpp::plogis(lvec);
+    
+    for (int i=0; i<N; i++) {
+      y(i) = as<double>(rbinom(1, 1, ps(i)));
+    }
+    
+    aStat = estAUC(dmatrix, y);
+    
+    // Rcout << auc << " " << alpha << " " << aStat << "\n";
+    
+    if (aStat < auc) {
+      intLow = alpha;
+    } else {
+      intHigh = alpha;
+    }
+    
+  }
+  
+  alpha =  (intLow + intHigh)/2;
+  
+  for (int i = 0; i < coefs.length(); i++) {
+    avec(i) = alpha * coefs(i);
+  }
+  
+  lvec = mat_Mult(covmat, avec);
+  B0 = getBeta0(lvec, popPrev, tolerance);
+  
+  results(0) = B0;
+  results(1) = alpha;
+  
+  return(results);
   
 }